Add conversion from bytes to words with padding in CRC DMA version

+++ Sts1CobcSw/Periphery/HardwareCrc32.hpp
@@ -6,6 +6,6 @@ namespace sts1cobcsw::periphery
 {
 auto EnableHardwareCrc() -> void;
 auto EnableHardwareCrcAndDma() -> void;
-auto DmaCrc32(std::span<uint32_t> data) -> void;
+auto DmaCrc32(std::span<uint8_t> data) -> void;
 auto HardwareCrc32(std::span<uint32_t> data) -> uint32_t;
 }  // namespace sts1cobcsw::periphery
